pyproject.toml
[project]
name = "test-cosmic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==5.2",
    "django-admin-sortable2>=2.3.1",
    "django-environ>=0.13.0",
    "django-filer>=3.4.4",
    "django-unfold>=0.81.0",
    "mysqlclient>=2.2.8",
]

[dependency-groups]
dev = [
    "gunicorn>=25.1.0",
]
